[MODIFY] separate meal and event module
이벤트와 급식 모듈 분리
